backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
diffusers
transformers
accelerate
scipy
ftfy
pillow
numpy